hw: split pc.c file
The hw/i386/pc.c file is almost for pc/q35 platform and
split it to pc.c and x86.c. x86.c is common file for all
x86 platform. Change the CONFIG definition for pc.c and
compile out this file if we build virt platform.
