+++ default-configs/i386-softmmu.mak
@@ -72,7 +72,6 @@ CONFIG_PCI_LITE=y
 CONFIG_ACPI_HW_REDUCED=y
 CONFIG_ACPI_PCI_HOTPLUG=y
 CONFIG_MULTIBOOT=y
-CONFIG_X86_COMMON=y
 CONFIG_X86_PIIX=y
 CONFIG_X86_Q35=y
 CONFIG_X86_VIRT=y
