+++ default-configs/x86_64-softmmu.mak
@@ -73,7 +73,6 @@ CONFIG_ACPI_X86_PIXX=y
 CONFIG_ACPI_HW_REDUCED=y
 CONFIG_ACPI_PCI_HOTPLUG=y
 CONFIG_MULTIBOOT=y
-CONFIG_X86_COMMON=y
 CONFIG_X86_PIIX=y
 CONFIG_X86_Q35=y
 CONFIG_X86_VIRT=y
