+++ hw/i386/Makefile.objs
@@ -1,7 +1,6 @@
 obj-$(CONFIG_KVM) += kvm/
 obj-$(CONFIG_MULTIBOOT) += multiboot.o
-obj-y += cpu.o fw.o memory.o kernel-loader.o
-obj-$(CONFIG_X86_COMMON) += pc.o
+obj-y += cpu.o x86.o fw.o memory.o kernel-loader.o
 obj-$(CONFIG_X86_PIIX) += pc_piix.o
 obj-$(CONFIG_X86_Q35) += pc_q35.o
 obj-y += sysfw.o
@@ -13,4 +12,8 @@ obj-$(CONFIG_VMPORT) += vmport.o
 obj-$(CONFIG_VMMOUSE) += vmmouse.o
 
 obj-$(CONFIG_KVM_VAPIC) += kvmvapic.o
-obj-$(call lor,$(CONFIG_X86_PIIX),$(CONFIG_X86_Q35)) += acpi-build.o
+
+ifeq ($(call lor,$(CONFIG_X86_PIIX),$(CONFIG_X86_Q35)),y)
+obj-y += acpi-build.o
+obj-y += pc.o
+endif
